feat: 글로벌 API 예외 처리기 등록
